pyproject.toml
[project]
name = "amps-m3u"                         # Your PyPI package name
version = "0.0.2"
description = "Dynamic M3U generator and FFmpeg-based streaming server."
readme = "README.md"
requires-python = ">=3.7"
license = { file = "LICENSE" }
authors = [
  { name="ChippyTech (CalmhostAcct)", email="chippymeows@gmail.com" }
]

dependencies = [
    "Flask",
    "flask-sock",
    "ffmpeg-python",
    "PyYAML",
    "yt-dlp",
    "click",
    "uvicorn",
    "apscheduler"
]

# Optional classifiers for PyPI
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

# The CLI entry point
[project.scripts]
amps = "amps.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["amps/**"]
